feat: Add ScrollTriggeredDiv component for scroll-triggered animations
feat: Implement FeaturesSection with animated feature cards

feat: Create FinalCTASection with animated call-to-action elements

feat: Develop Footer component for site navigation and information

feat: Build Header component with navigation and branding

feat: Construct HeroSection with engaging introductory content

feat: Add PricingSection for plan selection and details

feat: Create TestimonialsSection to showcase user feedback

chore: Define constants for landing page features and testimonials
