+++ frontend/src/features/layout/components/index.ts
@@ -1,1 +1,8 @@
-
+export { default as CustomPricingCards } from "./CustomPricingCards";
+export { default as Header } from "./Header";
+export { default as HeroSection } from "./HeroSection";
+export { default as TestimonialsSection } from "./TestimonialsSection";
+export { default as PricingSection } from "./PricingSection";
+export { default as FeaturesSection } from "./FeaturesSection";
+export { default as FinalCTASection } from "./FinalCTASection";
+export { default as Footer } from "./Footer";
