+++ frontend/src/features/layout/utils/landing-page-const.ts
@@ -1,0 +1,78 @@
+import {
+  PlusCircleIcon,
+  BarChartIcon,
+  TrendingUpIcon,
+  CheckIcon,
+} from "@/components/Icons";
+
+export const trustIndicators = [
+  {
+    icon: CheckIcon,
+    title: "Easy Setup",
+    description: "Get started in under 2 minutes",
+    color: "green",
+  },
+  {
+    icon: CheckIcon,
+    title: "Real-time Tracking",
+    description: "Instant macro calculations",
+    color: "blue",
+  },
+  {
+    icon: CheckIcon,
+    title: "Advanced Analytics",
+    description: "Deep insights & trends",
+    color: "purple",
+  },
+];
+
+export const trustElements = [
+  { text: "Always Free Version" },
+  { text: "No Credit Card Required" },
+  { text: "Cancel Anytime" },
+];
+
+export const features = [
+  {
+    icon: PlusCircleIcon,
+    title: "Smart Meal Logging",
+    description:
+      "Quickly log meals with our intelligent food database and barcode scanner. Get accurate macro breakdowns instantly.",
+  },
+  {
+    icon: BarChartIcon,
+    title: "Personal Goal Setting",
+    description:
+      "Set personalized macro targets based on your goals, activity level, and preferences. Track progress in real-time.",
+  },
+  {
+    icon: TrendingUpIcon,
+    title: "Advanced Analytics",
+    description:
+      "Visualize your progress with detailed charts, trends, and insights that help you optimize your nutrition strategy.",
+  },
+];
+
+export const testimonials = [
+  {
+    name: "Alex Peterson",
+    title: "Fitness Enthusiast",
+    quote:
+      "MacroTrackr completely transformed my approach to nutrition. The interface is intuitive, and the analytics help me stay consistent with my goals.",
+    initials: "AP",
+  },
+  {
+    name: "Jamie Lewis",
+    title: "Certified Nutrition Coach",
+    quote:
+      "As a nutrition coach, I recommend MacroTrackr to all my clients. It's the perfect balance of simplicity and powerful features.",
+    initials: "JL",
+  },
+  {
+    name: "Maria Rodriguez",
+    title: "Busy Professional",
+    quote:
+      "I've tried many tracking apps, but MacroTrackr is by far the most user-friendly. Lost 20 pounds and gained so much energy!",
+    initials: "MR",
+  },
+];
